Refactor MustSupportTest and MSChecker to streamline slice presence checks by moving related methods to a new Slices module, enhancing code organization and maintainability.
